Crate database_cli

About the project

This project was done for a high school presentation by me, Adrian Struwe (Eskaan). It is currently licenced under the GPL3 licence.

Inspiration and Credit

The idea for this project is not my own. This has ben done before by David Kiesel and was presented by him in the C36C3 Congress. Credit also goes to Marudor and his bahn.expert project, it was very helpful for me to understand the inner workings of HAFAS.

Project structure

The project is split into two main parts:

hafas_wrap, the library to access the HAFAS (Hacon Fahrplan Auskunfts System). This library contains the request and response data structures and a request system.

database_cli that uses data queried through the library and injects certain parts into a postgresql database for further processig. It also (currently) contains the subcommands to print a diagram showing the most used stations in the datbase. If you are searching about how to use this project, you will have to look here.

Each of those projects contain their own README.md file further explaining their use.

I also recommend using the rust-doc documentation for the project-level documentation, found in the doc folder (./doc/database_cli/index.html).

Goals

The formulated goal for this project was to query and compare train schedule to realtime data. Sadly, after many experiments I realized this API was designed to resist such data collection attempts. I am still trying to find a feasible way (that does not overload the API) to collect such data. As I could not finish with my targeted goal, I insted selected one more realistic; to display the most used stations in the database.

Future of the project

I plan on continuing to work on this and maybe one day make a good HAFAS library for Rust. This is the primary goal for the future, after which I plan to create a public API that serves better readable responses than HAFAS.

Database-cli toolchain

This is a command line interface using the hafas_wrap library.

Use and Usage

The current use is to create a database of trips, train_types, locations and operators. For the exact layout, see database_layout.png at the root of the project. You can also create a simple diagram from usage statistics.

Altrough the program does not need a live database to compile, it will need a postgres database to function. The database needs to be on localhost, with the name db-statistics and the user postgres (password-less) being able to access all tables specified in the database layout. To prepare the database for use, run database-cli migrate. This will create all neccecary schemas, tables, types and functions.

For the exact usage, refer to the dcumentation of the main method.

Modules

compare_raw_data 🔒

Functions for the cli sub-command data check

count_location_trips 🔒

Functions for the cli sub-command data parse_heatmap

create_heatmap_diagram 🔒

Functions for the cli sub-command create_heatmap

parse_raw_jids 🔒

Functions for the cli sub-command data parse

request_raw_jids 🔒

Misc functions to request raw HAFAS journey details.

util

Some utilities that are used program-wide.

Functions

main 🔒

The main method of the cli program.