Module database_cli::count_location_trips

Functions for the cli sub-command data parse_heatmap

Functions

get_eva_name

Fetches the name for a eva from the locations table.

increment_count

Increments the count of a specific entry in the location_counts table by count.

insert_or_nothing

Tries to insert a new entry into the location_counts table, does nothing if the primary key already exists.

parse_count

This function is called by the cli sub-command data parse_heatmap.