Function database_cli::create_heatmap_diagram::create_heatmap

pub async fn create_heatmap(
    output_file: Option<&String>,
    filter_cat_out: &String,
    filter_cat_code: &u8,
    limit_search: &i64,
    limit_entries: &usize
)

This function is called by the cli sub-command create_heatmap.

It creates a svg diagram from the data in the location_counts table in the database. The diagram can be filtered by a regex for cat_out and a smaller-or-equal-than clause for cat_code. See draw_diagram about the diagram specifics.