Function database_cli::main

pub(crate) fn main()

The main method of the cli program.

This method parses the system-given command arguments using the clap crate and calls the corresponding method.

List of sub-commands:

All commands can be passed -v or -vv to make them more verbose. You can add the --help argument to any subcommand to see details about the arguments. If you do not want to compile the cli, refer to the source code for arguments.

data

request_raw: This command requests raw schedule data over all jids in the HAFAS endpoint. Independenty of the TO argument, the command will panic if it reaches the last jid. The ususal last jid is around 1.5 Million.

It is recommended to leave all optional arguments to a default value to prevent a timeout.

parse: This command parses the data from the raw_data table to a usable format and insertis it into the other tables. It can also be called automatically by adding --parse to the arguments of request_raw.

parse_heatmap: This command is a command that should be called before using the create_heatmap feature. It counts together all of the recorded train trips into their own table for faster access. As near and local trafic can obfuscate the ending image, I recommend setting -o 'ICE' as filter.

check: This command checks if data from the HAFAS endpoint differs from the current data. This check is only done for a single jid. A difference might hint at a schedule change.

create_heatmap Creates a horizontal bar diagram of the most used stations in the lookup table. It can be filtered by cat_code, cat_out and search limit. For currently unkown reasons, anything over 11 Bars will mess up the station names.

migrate: Creates all neccessary infrastructure on the remote database.