Module database_cli::parse_raw_jids

Functions for the cli sub-command data parse

Structs

OpDays

Database type operation_dates.

ScheduledStop

Database type scheduled_stop

Functions

hex_to_binary 🔒

Parses a hexadecimal char into a four-digit binary String.

parse

This function is called by the cli sub-command data parse.

parse_chunk 🔒

Queries, parses a single chunk and inserts it back into the database.

parse_date 🔒

Parses a date from HAFAS into a [time::Date].

parse_entry

Parses a single serde_json::Value into a usable format and inserts it into the database.

parse_op_days 🔒

Parses the obfuscated String of operation days from HAFAS into a Vec of [time::Date].

parse_time 🔒

Parses a time from HAFAS into a [time::Time]